+++ aspen_builder/lib/src/bundle_generator.dart
@@ -1,7 +1,11 @@
+import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
+import 'package:aspen_assets/aspen_assets.dart';
 import 'package:build/build.dart';
+import 'package:glob/glob.dart';
 import 'package:source_gen/source_gen.dart';
+import 'package:path/path.dart' as p;
 
 import 'package:aspen/aspen.dart';
 
@@ -45,7 +49,17 @@ class BundleGenerator extends GeneratorForAnnotation<Asset> {
           '@Asset(...) ${variableElement.displayName} has an invalid type');
     }
 
-    var current = type as InterfaceType;
+    final loadableAsset = getLoadableAssetForType(type);
+
+    if (loadableAsset == null) {
+      error(variableElement,
+          'Asset type ${type.getDisplayString(withNullability: true)} has no @LoadableAsset annotation');
+    }
+    return loadableAsset;
+  }
+
+  LoadableAsset getLoadableAssetForType(InterfaceType type) {
+    var current = type;
     while (current != null) {
       for (var ann in current.element.metadata) {
         var annValue = ann.computeConstantValue();
@@ -58,9 +72,6 @@ class BundleGenerator extends GeneratorForAnnotation<Asset> {
 
       current = current.superclass;
     }
-
-    error(variableElement,
-        'Asset type ${type.getDisplayString()} has no @LoadableAsset annotation');
     return null;
   }
 
@@ -80,32 +91,87 @@ class BundleGenerator extends GeneratorForAnnotation<Asset> {
       return Future.value();
     }
 
-    var loadableAsset = getLoadableAsset(variableElement);
+    final assetPathReader = _getAssetPathReader(annotation);
+    final assetPath = assetPathReader.stringValue;
+
+    if (variableElement.type.element.name == 'DirAsset') {
+      final assetDirId = AssetId.resolve(assetPath, from: buildStep.inputId);
+      if (buildStep.inputId.package != assetDirId.package) {
+        error(element,
+            'DirAsset only works for the current package (current: ${buildStep.inputId.package}, requested: ${assetDirId.package})');
+      }
+      var assetDirPath = assetDirId.path;
+      final firstIndexOfWildcard = assetDirPath.indexOf('*');
+      if (firstIndexOfWildcard == -1) {
+        assetDirPath = '$assetDirPath/*.*';
+      }
+
+      final assetType =
+          (variableElement.type as InterfaceType).typeArguments[0];
+      final loadableAsset = getLoadableAssetForType(assetType);
+
+      if (loadableAsset == null) {
+        error(variableElement,
+            'DirAsset type $assetType has no @LoadableAsset annotation');
+        return Future.value();
+      }
+
+      final enumName =
+          '${element.name[0].toUpperCase()}${element.name.substring(1)}';
+      final enumValues = <String, AssetData>{};
+      final enumValueSanitzerRegExp = RegExp('[^a-zA-Z0-9]');
+
+      final assetIds = buildStep.findAssets(Glob(assetDirPath));
+      await for (final assetId in assetIds) {
+        var enumValue =
+            '${p.basenameWithoutExtension(assetId.path).replaceAll(enumValueSanitzerRegExp, '')}';
+        final extension = p.extension(assetId.path);
+        if (extension.isNotEmpty) {
+          enumValue =
+              '$enumValue\$${extension.replaceAll(enumValueSanitzerRegExp, '')}';
+        }
+        enumValues[enumValue] = await _loadAssetData(
+            loadableAsset, buildStep, assetId, element, value);
+      }
+
+      return Future.value('''
+enum $enumName {
+  ${enumValues.keys.join(', ')}
+}
+const _${variableElement.name}\$asset = {
+  ${enumValues.entries.map((entry) => '$enumName.${entry.key}: ${assetType.element.name}(AssetData(r\'${entry.value.assetId}\', r\'\'\'${entry.value.content}\'\'\'))').join(',\n')}
+};
+''');
+    } else {
+      final assetId = AssetId.resolve(assetPath, from: buildStep.inputId);
+      final loadableAsset = getLoadableAsset(variableElement);
+      final assetData = await _loadAssetData(
+          loadableAsset, buildStep, assetId, element, value);
+
+      return Future.value(
+          "const _${variableElement.name}\$asset = AssetData(r'${assetData.assetId}', r'''${assetData.content}''');");
+    }
+  }
+
+  Future<AssetData> _loadAssetData(
+      LoadableAsset loadableAsset,
+      BuildStep buildStep,
+      AssetId assetId,
+      Element element,
+      DartObject value) async {
+    final assetIdString = assetId.toString();
     if (loadableAsset == null) {
-      return Future.value();
+      return AssetData(assetIdString, '');
     }
 
     assert(
         loadableAsset.url == 'package:aspen_builder/src/default_loaders.dart');
 
-    var loader = defaultLoaders[loadableAsset.loader];
+    final loader = defaultLoaders[loadableAsset.loader];
 
-    var assetPathReader;
-
-    if (options.config['release_asset'] == true) {
-      assetPathReader = annotation.read('releasePath');
-      if (assetPathReader.isNull) {
-        assetPathReader = annotation.read('path');
-      }
-    } else {
-      assetPathReader = annotation.read('path');
-    }
-
-    var assetId =
-        AssetId.resolve(assetPathReader.stringValue, from: buildStep.inputId);
     if (!await buildStep.canRead(assetId)) {
       error(element, 'Asset ${assetId} cannot be found');
-      return Future.value();
+      return AssetData(assetIdString, '');
     }
 
     String content;
@@ -121,10 +187,22 @@ class BundleGenerator extends GeneratorForAnnotation<Asset> {
         error(element, '${loadableAsset.loader} failed with: ${ex.message}');
       }
 
-      return Future.value();
+      return AssetData(assetIdString, '');
     }
+    return AssetData(assetIdString, content);
+  }
 
-    return Future.value(
-        "const _${variableElement.name}\$asset = AssetData(r'$assetId', r'''${content}''');");
+  ConstantReader _getAssetPathReader(ConstantReader annotation) {
+    var assetPathReader;
+
+    if (options.config['release_asset'] == true) {
+      assetPathReader = annotation.read('releasePath');
+      if (assetPathReader.isNull) {
+        assetPathReader = annotation.read('path');
+      }
+    } else {
+      assetPathReader = annotation.read('path');
+    }
+    return assetPathReader;
   }
 }