+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - the declaration of an asset changed (for all types of assets):
     - 0.3.0: `const myAsset = TextAsset(text: _myAsset$content);` 
     - 0.4.0: `const myAsset = TextAsset(_myAsset$asset);` 
+- added new asset type `DirAsset` which adds all (or selected) files in a given directory
 
 # 0.3.0